MEM: move declaration of multidim classes to a header and make them non-final
